Adicionado Sugestões

+++ src/main/java/br/edu/ifrs/canoas/webapp/domain/Score.java
@@ -25,5 +25,5 @@ public abstract class Score {
     private String label;
 
     @NotNull
-    private Float weight;
+    private double weight;
 }
